+++ CHANGELOG.md
@@ -1,4 +1,11 @@
 # Changelog
+## [0.4.7]
+- add csv
+- add jq
+- add to_csv/to_json/to_toml
+- fix err msg for parsing module.
+- prompt settings.
+
 ## [0.4.6]
 - refactor sort function
 - optimize filesize, time display. optimize list.group.
@@ -6,7 +13,9 @@
 - fix 1M in line end.
 - fix non-ascii path broken
 - add ..., suport justify range start via +/-; fix fs.mv/cp
-- more detail while load history.
+- fix history.
+- add vi mode.
+
 
 ## [0.4.5]
 - fix stdin read for third cmd